+++ doc/release_notes.rst
@@ -35,7 +35,7 @@ Upcoming Release
 
 * Add support for the linopy ``io_api`` option; set to ``"direct"`` to increase model reading and writing performance for the highs and gurobi solvers.
 
-* Add several map projection options for plotting.
+* Add the option to customise map projection in plotting config.
 
 
 PyPSA-Eur 0.9.0 (5th January 2024)